Move stellarators and Structures scenarios
