--- inductiva/stellarators/scenarios/coils/__init__.py
@@ -1,2 +1,0 @@
-#pylint: disable=missing-module-docstring
-from .coils import StellaratorCoils
