+++ inductiva/structures/__init__.py
@@ -1,5 +1,5 @@
 #pylint: disable=missing-module-docstring
-from .scenarios import DeformablePlate
+from .plate_linear_elastic import DeformablePlate
 from . import bcs
 from . import holes
 from . import materials
