+++ inductiva/structures/plate_linear_elastic/__init__.py
