+++ inductiva/structures/plate_linear_elastic/bcs_utils.py
