+++ inductiva/structures/plate_linear_elastic/geometry_utils.py
