+++ inductiva/structures/plate_linear_elastic/plate_linear_elastic.py
