--- inductiva/structures/scenarios/plate_linear_elastic/__init__.py
@@ -1,2 +1,0 @@
-# pylint: disable=missing-module-docstring
-from .plate_linear_elastic import DeformablePlate
